finished here

+++ src/shared-kanban-board/shared-kanban-board.service.ts
@@ -2,14 +2,16 @@ import { Inject, Injectable } from '@nestjs/common';
 import { REPOSITORIES } from 'src/common/constants';
 import { SharedKanbanBoard } from './shared-kanban-board.model';
 import { FindOptions } from 'sequelize';
+import { SharedKanbanBoardDto } from './dtos/create-shared-kanban-board.dto';
+import { PartialSharedKanbanBoardDto } from './dtos/partial-shared-kanban-board.dto';
 
 @Injectable()
 export class SharedKanbanBoardService {
     constructor(@Inject(REPOSITORIES.SHARED_KANBAN_REPOSITORY)
     private sharedModel: typeof SharedKanbanBoard) { }
 
-    create(kId: number, uId: number) {
-        return this.sharedModel.create({ kanbanId: kId, userId: uId });
+    create(createDto: SharedKanbanBoardDto) {
+        return this.sharedModel.create({ kanbanId: createDto.kanbanId, userId: createDto.userId });
     }
 
     async get(userId: number) {
@@ -21,11 +23,11 @@ export class SharedKanbanBoardService {
         return this.sharedModel.findAll(options);
     }
 
-    findOne(options:FindOptions) {
+    findOne(options: FindOptions) {
         return this.sharedModel.findOne(options);
     }
 
-    delete(userId: number, kanbanId: number) {
-        return this.sharedModel.destroy({ where: { userId, kanbanId } })
+    delete(dto: SharedKanbanBoardDto) {
+        return this.sharedModel.destroy({ where: { userId: dto.userId, kanbanId: dto.kanbanId } })
     }
 }
